+++ juno.config.ts
@@ -1,4 +1,5 @@
 import { defineConfig } from "@junobuild/config";
+import { REDIRECTS } from "./redirects.ts";
 
 export default defineConfig({
   satellite: {
@@ -9,13 +10,11 @@ export default defineConfig({
     predeploy: ["deno task docs:build", "deno task predeploy"],
     source: "dist",
     storage: {
-      redirects: [
-        {
-          source: "/",
-          location: "/core/latest",
-          code: 302,
-        },
-      ],
+      redirects: REDIRECTS,
+    },
+    precompress: {
+      mode: "replace", // only upload the compressed version of the file
+      algorithm: "brotli",
     },
   },
   emulator: {